fix(platform): fix more platform-specific type issues
Fixed remaining macOS compilation errors:

**device.rs:**
- Added #[allow(clippy::unnecessary_cast)] to S_IFBLK and S_IFSOCK
- Same cross-platform issue: SFlag::bits() returns different types
  on different platforms

**metadata.rs:**
- Changed statx_impl from #[cfg(unix)] to #[cfg(target_os = "linux")]
- statx syscall is Linux-specific, not available on macOS
- CString only needed by Linux io_uring statx implementation

**directory.rs:**
- Changed DirectoryFd::statx() from #[cfg(unix)] to #[cfg(target_os = "linux")]
- Since statx_impl is Linux-only, the method must be too

**symlink.rs:**
- Fixed nix::unistd guard to #[cfg(all(unix, not(target_os = "linux")))]
- Used by macOS symlinkat, not on Linux

Fixes macOS errors:
- mismatched types (u16 | u32)
- failed to resolve: use of undeclared type CString
- unused import: nix::unistd

+++ crates/compio-fs-extended/src/device.rs
@@ -194,7 +194,9 @@ pub async fn create_block_device_at_path(
     let dev = ((major as u64 & 0xfff) << 8)
         | (minor as u64 & 0xff)
         | (((major as u64 >> 12) & 0xfffff) << 32);
-    let device_mode = stat::SFlag::S_IFBLK.bits() | (mode & 0o777);
+    // Note: On some platforms SFlag::bits() returns u16, on others u32
+    #[allow(clippy::unnecessary_cast)]
+    let device_mode = stat::SFlag::S_IFBLK.bits() as u32 | (mode & 0o777);
 
     create_special_file_at_path(path, device_mode, dev).await
 }
@@ -230,7 +232,9 @@ pub async fn create_block_device_at_path(
 /// - The operation fails due to I/O errors
 #[cfg(unix)]
 pub async fn create_socket_at_path(path: &Path, mode: u32) -> Result<()> {
-    let socket_mode = stat::SFlag::S_IFSOCK.bits() | (mode & 0o777);
+    // Note: On some platforms SFlag::bits() returns u16, on others u32
+    #[allow(clippy::unnecessary_cast)]
+    let socket_mode = stat::SFlag::S_IFSOCK.bits() as u32 | (mode & 0o777);
 
     create_special_file_at_path(path, socket_mode, 0).await
 }
